+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 ### Added
 - PixPullSubscription resource
 - PixPullRequest resource
+- IssuingStockRule resource
 
 ## [0.18.0] - 2026-06-22
 ### Changed